requirements.txt
Django==1.8.3
django-multiselectfield==0.1.3
djangorestframework==3.1.3
